Crate kafka

Clients for comunicating with a Kafka cluster. These are:

kafka::producer::Producer - for sending message to Kafka

kafka::consumer::Consumer - for retrieving/consuming messages from Kafka

kafka::client::KafkaClient - a lower-level, general purpose client leaving you with more power but also more resposibility

See module level documentation corresponding to each client individually.

Reexports

pub use self::error::{Error, Result};

Modules

client
Kafka Client - A mid-level abstraction for a kafka cluster allowing building higher level constructs.

consumer
Kafka Consumer - A higher-level API for consuming kafka topics.

error
Error struct and methods

producer
Kafka Producer - A higher-level API for sending messages to Kafka topics.