Module kafka::client

Kafka Client - A mid-level abstraction for a kafka cluster allowing building higher level constructs.

The entry point into this module is KafkaClient obtained by a call to KafkaClient::new().

Modules

fetch
A representation of fetched messages from Kafka.

metadata
Types related to topic metadata for introspection by clients. Example: KafkaClient::topics().

Structs

CommitOffset
Data point identifying a particular topic partition offset to be commited. See KafkaClient::commit_offsets.

FetchGroupOffset
Data point identifying a topic partition to fetch a group's offset for. See KafkaClient::fetch_group_offsets.

FetchPartition
Partition related request data for fetching messages. See KafkaClient::fetch_messages.

KafkaClient
Client struct keeping track of brokers and topic metadata.

PartitionOffset
A retrieved offset for a particular partition in the context of an already known topic.

ProduceMessage
Message data to be sent/produced to a particular topic partition. See KafkaClient::produce_messages and Producer::send.

SecurityConfig
Security relevant configuration options for KafkaClient.

TopicPartitionOffset
A retrieved offset of a particular topic partition.

Enums

Compression
Compression types supported by kafka. The numeral values of this enumeration correspond to the compression encoding in the attributes of a Message in the protocol.

FetchOffset
Possible values when querying a topic's offset. See KafkaClient::fetch_offsets.

GroupOffsetStorage
Defines the availale storage types to utilize when fetching or comitting group offsets. See also KafkaClient::set_group_offset_storage.

RequiredAcks
Possible choices on acknowledgement requirements when producing/sending messages to Kafka. See KafkaClient::produce_messages.

Constants

DEFAULT_COMPRESSION
The default value for KafkaClient::set_compression(..)

DEFAULT_CONNECTION_IDLE_TIMEOUT_MILLIS
The default value for KafkaClient::set_connection_idle_timeout(..)

DEFAULT_FETCH_CRC_VALIDATION
The default value for KafkaClient::set_fetch_crc_validation(..)

DEFAULT_FETCH_MAX_BYTES_PER_PARTITION
The default value for KafkaClient::set_fetch_max_bytes(..)

DEFAULT_FETCH_MAX_WAIT_TIME_MILLIS
The default value for KafkaClient::set_fetch_max_wait_time(..)

DEFAULT_FETCH_MIN_BYTES
The default value for KafkaClient::set_fetch_min_bytes(..)

DEFAULT_GROUP_OFFSET_STORAGE
The default value for KafkaClient::set_group_offset_storage(..)

DEFAULT_RETRY_BACKOFF_TIME_MILLIS
The default value for KafkaClient::set_retry_backoff_time(..)

DEFAULT_RETRY_MAX_ATTEMPTS
The default value for KafkaClient::set_retry_max_attempts(..)