Struct kafka::client::fetch::Data

pub struct Data<'a> {
    // some fields omitted
}

The successfully fetched data payload for a particular partition.

Methods

impl<'a> Data<'a>

fn highwatermark_offset(&self) -> i64

Retrieves the so-called "high water mark offset" indicating the "latest" offset for this partition at the remote broker. This can be used by clients to find out how much behind the latest available message they are.

fn messages(&self) -> &[Message<'a>]

Retrieves the fetched message data for this partition.