Struct kafka::client::fetch::Partition

pub struct Partition<'a> {
    // some fields omitted
}

The result of a "fetch messages" request from a particular Kafka broker for a single topic partition only. Beside the partition identifier, this structure provides an iterator over the actually requested message data.

Note: There might have been a (recoverable) error for a particular partition (but not for another).

Methods

impl<'a> Partition<'a>

fn partition(&self) -> i32

Retrieves the identifier of the represented partition.

fn data(&'a self) -> &'a Result<Data<'a>>

Retrieves the data payload for this partition.