Struct kafka::client::metadata::Partition

pub struct Partition<'a> {
    // some fields omitted
}

Metadata about a particular topic partition.

A partition can be seen as either available or not by kafka-rust. "Available" partitions are partitions with an assigned leader broker and can be send messages to or fetched messages from. Non-available partitions are ignored by kafka-rust. Whether or not a partition is currently "available" can be determined by testing for partition.leader().is_some() or more directly through partition.is_available().

Methods

impl<'a> Partition<'a>

fn id(&self) -> i32

Retrieves the identifier of this topic partition.

fn leader(&self) -> Option<&'a Broker>

Retrieves the current leader broker of this partition - if any. A partition with a leader is said to be "available".

fn is_available(&self) -> bool

Determines whether this partition is currently "available". See Partition::leader().