Struct kafka::client::metadata::Partitions

pub struct Partitions<'a> {
    // some fields omitted
}

Metadata relevant to partitions of a particular topic.

Methods

impl<'a> Partitions<'a>

fn len(&self) -> usize

Retrieves the number of the topic's partitions.

fn is_empty(&self) -> bool

Tests for .len() > 0.

fn iter(&self) -> PartitionIter<'a>

Retrieves an iterator of the partitions of the underlying topic.

fn partition(&self, partition_id: i32) -> Option<Partition<'a>>

Finds a specified partition identified by its id.

fn available_ids(&self) -> Vec<i32>

Convenience method to retrieve the identifiers of all currently "available" partitions. Such partitions are known to have a leader broker and can be sent messages to.