Struct kafka::client::metadata::Topics

pub struct Topics<'a> {
    // some fields omitted
}

A view on the loaded metadata about topics and their partitions.

Methods

impl<'a> Topics<'a>

fn new(client: &KafkaClient) -> Topics

Constructs a view of the currently loaded topic metadata from the specified kafka client.

fn len(&self) -> usize

Retrieves the number of the underlying topics.

fn iter(&'a self) -> TopicIter<'a>

Provides an iterator over the underlying topics.

fn names(&'a self) -> TopicNames<'a>

A conveniece method to return an iterator over the topics' names.

fn contains(&'a self, topic: &str) -> bool

A convenience method to determine whether the specified topic is known.

fn partitions(&'a self, topic: &str) -> Option<Partitions<'a>>

Retrieves the partitions of a specified topic.