Struct kafka::consumer::Consumer

pub struct Consumer {
    // some fields omitted
}

The Kafka Consumer

See module level documentation.

Methods

impl Consumer

fn from_client(client: KafkaClient) -> Builder

Starts building a consumer using the given kafka client.

fn from_hosts(hosts: Vec<String>) -> Builder

Starts building a consumer bootstraping internally a new kafka client from the given kafka hosts.

fn client(self) -> KafkaClient

Destroys this consumer returning back the underlying kafka client.

fn poll(&mut self) -> Result<MessageSets>

Polls for the next available message data.

fn group(&self) -> &str

Retrieves the group on which behalf this consumer is acting. The empty group name specifies a group-less consumer.

fn last_consumed_message(&self, topic: &str, partition: i32) -> Option<i64>

Retrieves the offset of the last "consumed" message in the specified partition. Results in None if there is no such "consumed" message.

fn consume_message(&mut self, topic: &str, partition: i32, offset: i64) -> Result<()>

Marks the message at the specified offset in the specified topic partition as consumed by the caller.

Note: a message with a "later/higher" offset automatically marks all preceeding messages as "consumed", this is messages with "earlier/lower" offsets in the same partition. Therefore, it is not neccessary to invoke this method for every consumed message.

Results in an error if the specified topic partition is not being consumed by this consumer.

fn consume_messageset<'a>(&mut self, msgs: MessageSet<'a>) -> Result<()>

A convience method to mark the given message set consumed as a whole by the caller. This is equivalent to marking the last message of the given set as consumed.

fn commit_consumed(&mut self) -> Result<()>

Persists the so-far "marked as consumed" messages (on behalf of this consumer's group for the underlying topic - if any.)

See also Consumer::consume_message and Consumer::consume_messetset.