Struct kafka::consumer::Message

pub struct Message<'a> {
    pub offset: i64,
    pub key: &'a [u8],
    pub value: &'a [u8],
}

A fetched message from a remote Kafka broker for a particular topic partition.

Fields
offset: i64

The offset at which this message resides in the remote kafka broker topic partition.

key: &'a [u8]

The "key" data of this message. Empty if there is no such data for this message.

value: &'a [u8]

The value data of this message. Empty if there is no such data for this message.