Struct kafka::consumer::MessageSets

pub struct MessageSets {
    // some fields omitted
}

Messages retrieved from kafka in one fetch request. This is a concatenation of blocks of messages successfully retrieved from the consumed topic partitions. Each such partitions is guaranteed to be present at most once in this structure.

Methods

impl MessageSets

fn is_empty(&self) -> bool

Determines efficiently whether there are any consumeable messages in this data set.

fn iter(&self) -> MessageSetsIter

Iterates over the message sets delivering the fetched message data of consumed topic partitions.