Enum kafka::error::Error

pub enum Error {
    Io(Error),
    InvalidInputSnappy,
    UnexpectedEOF,
    CodecError,
    NoHostReachable,
    Unknown,
    OffsetOutOfRange,
    InvalidMessage,
    UnknownTopicOrPartition,
    InvalidMessageSize,
    LeaderNotAvailable,
    NotLeaderForPartition,
    RequestTimedOut,
    BrokerNotAvailable,
    ReplicaNotAvailable,
    MessageSizeTooLarge,
    StaleControllerEpochCode,
    OffsetMetadataTooLargeCode,
    OffsetsLoadInProgressCode,
    ConsumerCoordinatorNotAvailableCode,
    NotCoordinatorForConsumerCode,
}

The error type for various operations.

Extends io::Error with following

InvalidInputSnappy - Snappy failed to decode response from Kafka

UnexpectedEOF - Raised from byteorder when decoding values fails or upon the deliver of an insufficient number of bytes.

NoHostReachable - Unable to reach any host.

CodecError - Error while encoding or decoding request/response

Kafka Errors from https://cwiki.apache.org/confluence/display/KAFKA/A+Guide+To+The+Kafka+Protocol#AGuideToTheKafkaProtocol-ErrorCodes

Variants

Io

InvalidInputSnappy

UnexpectedEOF

CodecError

NoHostReachable

Unknown

OffsetOutOfRange

InvalidMessage

UnknownTopicOrPartition

InvalidMessageSize

LeaderNotAvailable

NotLeaderForPartition

RequestTimedOut

BrokerNotAvailable

ReplicaNotAvailable

MessageSizeTooLarge

StaleControllerEpochCode

OffsetMetadataTooLargeCode

OffsetsLoadInProgressCode

ConsumerCoordinatorNotAvailableCode

NotCoordinatorForConsumerCode