Struct kafka::producer::Partitions

pub struct Partitions {
    // some fields omitted
}

Producer relevant partition information of a particular topic.

Indented for use by Partitions.

Methods

impl Partitions

fn available_ids(&self) -> &[i32]

Retrieves the list of the identifiers of currently "available" partitions for the given topic. This list excludes partitions which do not have a leader broker assigned.

fn num_available(&self) -> u32

Retrieves the number of "available" partitions. This is a merely a convenience method. See Partitions::available_ids.

fn num_all(&self) -> u32

The total number of partitions of the underlygin topic. This number includes also partitions without a current leader assignment.