Struct kafka::producer::Record

pub struct Record<'a, K, V> {
    pub key: K,
    pub value: V,
    pub topic: &'a str,
    pub partition: i32,
}

A structure representing a message to be sent to Kafka through the Producer API. Such a message is basically a key/value pair specifying the target topic and optionally the topic's partition.

Fields
key: K

Key data of this (message) record.

value: V

Value data of this (message) record.

topic: &'a str

Name of the topic this message is supposed to be delivered to.

partition: i32

The partition id of the topic to deliver this message to. This partition may be < 0 in which case it is considered "unspecified". A Producer will then typically try to derive a partition on its own.

Methods

impl<'a, K, V> Record<'a, K, V>

fn from_key_value(topic: &'a str, key: K, value: V) -> Record<'a, K, V>

Convenience function to create a new key/value record with an "unspecified" partition - this is, a partition set to a negative value.

fn with_partition(self, partition: i32) -> Self

Convenience method to set the partition.

impl<'a, V> Record<'a, (), V>

fn from_value(topic: &'a str, value: V) -> Record<'a, (), V>

Convenience function to create a new value only record with an "unspecified" partition - this is, a partition set to a negative value.