Struct kafka::producer::Builder

pub struct Builder<P = DefaultPartitioner> {
    // some fields omitted
}

A Kafka Producer builder easing the process of setting up various configuration settings.

Methods

impl Builder

fn with_security(self, security: SecurityConfig) -> Self

Specifies the security config to use. See KafkaClient::new_secure for more info.

fn with_compression(self, compression: Compression) -> Self

Sets the compression algorithm to use when sending out data.

See KafkaClient::set_compression.

fn with_ack_timeout(self, timeout: i32) -> Self

Sets the maximum time in milliseconds the kafka brokers can await the receipt of required acknowledgements (which is specified through Builder::with_required_acks.) Note that Kafka explicitely documents this not to be a hard limit.

fn with_required_acks(self, required_acks: i16) -> Self

Sets how many acknowledgements the kafka brokers should receive before responding to sent messages. If it is 0 the servers will not send any response. If it is 1, the server will wait the data is written to the local server log before sending a replying. If it is -1 the servers will block until the messages are committed by all in sync replicas before replaying. For any number > 1 the servers will block waiting for this number of acknowledgements to occur (but the servers will never wait for more acknowledgements than there are in-sync replicas).

impl<P> Builder<P>

fn with_partitioner<Q: Partitioner>(self, partitioner: Q) -> Builder<Q>

Sets the partitioner to dispatch when sending messages without an explicit partition assignment.

fn create(self) -> Result<Producer<P>>

Finally creates/builds a new producer based on the so far supplied settings.